remove the apparently buggy box drawing API function
for some reason it's ok with the original "debug draw line" function,
but would randomly crash when an analogous "debug draw box" function
was called. so to be pragmatic i'll just try to roll all debug drawing
into this function with an extra `draw_type` argument to switch between
behaviors.